requirements.txt
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
zipp==3.8.0
psycopg2-binary
flask_httpauth
pillow
werkzeug
glob2
joblib
pandas
